split logic in libs
